Rename spec/ to test/ and group tests by concern
Moves the suite from a flat `spec/` directory to `test/` with
subdirectories that reflect what each test exercises:

- test/analysis/ — keyname/datatype analysis and depth/limit/sort behavior
- test/cli/ — bin/variety wrapper, option parsing, and mongosh translation
- test/persistence/ — varietyResults persistence and continuous logging
- test/plugins/ — plugin loader
- test/fixtures/ — shared static inputs (renamed from assets/)
- test/utils/ — Mocha helpers (Tester, MongoShell, JsonValidator)
- test/bin/ — the docker test harness

Updates imports, the typecheck/ESLint globs, package.json scripts, and
the CONTRIBUTING.md references to match. `test/fixtures/**` takes over
the shell-executed-fixture carve-out previously keyed on `spec/assets/**`.
