+++ test/analysis/ExcludeSubkeysTest.js
@@ -1,4 +1,4 @@
-import Tester from './utils/Tester.js';
+import Tester from '../utils/Tester.js';
 const test = new Tester('test', 'users');
 
 const sampleData = [
