+++ test/analysis/MaxDepthAnalysisTest.js
@@ -1,4 +1,4 @@
-import Tester from './utils/Tester.js';
+import Tester from '../utils/Tester.js';
 const test = new Tester('test', 'users');
 
 describe('Max-depth-limited analysis', () => {
