+++ test/fixtures/ExpectedAscii.js
