+++ test/persistence/ContinuousLoggingTest.js
@@ -1,6 +1,6 @@
 import assert from 'assert';
-import Tester from './utils/Tester.js';
-import sampleData from './assets/SampleData.js';
+import Tester from '../utils/Tester.js';
+import sampleData from '../fixtures/SampleData.js';
 
 const test = new Tester('test', 'users');
